+++ src/root.tsx
@@ -10,12 +10,12 @@ const Root = () => {
       <Routes>
         <Route path="/" element={<App />}>
           <Route index element={<HomePage />} />
-          <Route path="people/:person?">
+          <Route path="people/:slug?">
             <Route index element={<PeoplePage />} />
-            <Route path=":person" element={<PeoplePage />} />
+            <Route path=":slug" element={<PeoplePage />} />
           </Route>
           <Route path="*" element={<h1 className="title">Page not found</h1>} />
-          <Route path="home" element={<Navigate to="/" />} />
+          <Route path="home" element={<Navigate to="/" replace={true} />} />
         </Route>
       </Routes>
     </Router>
